requirements.txt
# Core Backend
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.4

# Data Processing (Excel datasets only)
pandas==2.2.2
openpyxl==3.1.2

# Multi-Agent + LLM
langchain==0.3.0
langchain-core==0.3.0
langchain-community==0.3.0
langchain-openai==0.1.7
langchain-anthropic==0.3.0
langgraph==0.2.0

openai==1.30.1
python-dotenv==1.0.1

# Frontend UI
streamlit==1.35.0

# API Tools
requests==2.31.0

# Voice + Multilingual (Best Quality)
gTTS==2.5.1
SpeechRecognition==3.10.1
pydub==0.25.1
faster-whisper==1.0.3
edge-tts==6.1.12
sounddevice==0.4.6
numpy==1.26.4

# Observability (Agent Tracing)
langsmith==0.1.147
